+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [3.4.0] - (17.11.2020)
+## Changs
+- added master Power control
+- fixed bug in mute control
+- update config.schema.json
 
 ## [3.3.1] - (14.09.2020)
 ## Changs